Implement custom user storage provider
The custom user storage provider was implemented to interact with an external service for user data retrieval and authentication. This involved creating user models and enhancing the Maven workflow for building and publishing the package. Specific Java classes `CustomUserStorageProvider`, `CustomUser`, `RespValidate`, `RespGetUser`, and others have been added or updated, as well as properties in `pom.xml`. A Maven workflow file was deleted in this process.